add new feature
the program runs inside a while loop that continues until the user decides to exit the program. After each check, the program asks the user if they want to check another word. If the user responds with anything other than "yes" (ignoring case), the loop will break and the program will terminate.